requirements.txt
selenium==3.7.0
PyYAML==3.12
pytest == 3.31